removed int64

+++ model/train.py
@@ -52,10 +52,10 @@ if __name__ == "__main__":
     os.mkdir(savepath)
 
     train_x = torch.load(args.input_path)[0:500]
-    train_y = torch.load(args.target_path).type(torch.int64)[0:500]
+    train_y = torch.load(args.target_path)[0:500]
 
     valid_x = torch.load(args.input_path)[500:]
-    valid_y = torch.load(args.target_path).type(torch.int64)[500:]
+    valid_y = torch.load(args.target_path)[500:]
 
     with open(savepath + "/setup.log", "w") as f: 
         f.write(run_name + "\n")
